distance between two complete

+++ db/migrations/20181121202312_locations.js
@@ -4,15 +4,13 @@ exports.up = function(knex, Promise) {
     table.integer('user_id').unsigned().notNull();
     table.unique('user_id');
     table.index('user_id');
-    table.float('lat', 14, 10);
-    table.float('long', 14, 10);
+    table.decimal('lat', 14 , 10);
+    table.decimal('long',14 , 10);
     table.dateTime('last_check_in');
 
     table.foreign('user_id').references('id').inTable('users');
   })
- 
  };
- 
  exports.down = function(knex, Promise) {
   return knex.schema.dropTable('locations');
  };
